fastapi/requirements.txt
fastapi
uvicorn[standard]
requests
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-multipart

